+++ nitrokey/src/auth.rs
@@ -149,7 +149,10 @@ where
     A: AuthenticatedDevice<D>,
     T: Fn(*const i8, *const i8) -> c_int,
 {
-    let temp_password = generate_password(TEMPORARY_PASSWORD_LENGTH);
+    let temp_password = match generate_password(TEMPORARY_PASSWORD_LENGTH) {
+        Ok(temp_password) => temp_password,
+        Err(err) => return Err((device, err)),
+    };
     let password = match get_cstring(password) {
         Ok(password) => password,
         Err(err) => return Err((device, err)),