+++ rand/utils/ci/script.sh
@@ -12,8 +12,10 @@ main() {
     cross test --target $TARGET --manifest-path rand_isaac/Cargo.toml --features=serde1
     cross test --target $TARGET --manifest-path rand_pcg/Cargo.toml --features=serde1
     cross test --target $TARGET --manifest-path rand_xorshift/Cargo.toml --features=serde1
+    cross test --target $TARGET --manifest-path rand_xoshiro/Cargo.toml
     cross test --target $TARGET --manifest-path rand_chacha/Cargo.toml
     cross test --target $TARGET --manifest-path rand_hc/Cargo.toml
+    cross test --target $TARGET --manifest-path rand_os/Cargo.toml
 }
 
 # we don't run the "test phase" when doing deploys